+++ renderer.js
@@ -2,7 +2,7 @@
 
 // Alpaca compatibility for Bootstrap Multiselect
 $.fn.extend({
-	multipleSelect: $.fn.multiselect
+	multipleSelect: $.fn.multiselect,
 });
 
 // Should we be in dark mode?
@@ -50,7 +50,7 @@ wwsumodules
 		host: machineID,
 		usernameField: "host",
 		authPath: "/auth/host",
-		authName: "Host"
+		authName: "Host",
 	})
 	.add("WWSUdirectors", WWSUdirectors, { host: machineID })
 	.add("directorReq", WWSUreq, {
@@ -59,7 +59,7 @@ wwsumodules
 		filter: null,
 		usernameField: "name",
 		authPath: "/auth/director",
-		authName: "Director"
+		authName: "Director",
 	})
 	.add("adminDirectorReq", WWSUreq, {
 		host: machineID,
@@ -67,7 +67,7 @@ wwsumodules
 		filter: { admin: true },
 		usernameField: "name",
 		authPath: "/auth/admin-director",
-		authName: "Administrator Director"
+		authName: "Administrator Director",
 	})
 	.add("masterDirectorReq", WWSUreq, {
 		host: machineID,
@@ -75,7 +75,7 @@ wwsumodules
 		filter: { ID: 1 },
 		usernameField: "name",
 		authPath: "/auth/admin-director",
-		authName: "Master Director"
+		authName: "Master Director",
 	})
 	.add("WWSUconfig", WWSUconfig)
 	.add("WWSUlogs", WWSUlogs)
@@ -86,7 +86,7 @@ wwsumodules
 		filter: { active: true },
 		usernameField: "name",
 		authPath: "/auth/dj",
-		authName: "DJ"
+		authName: "DJ",
 	})
 	.add("WWSUstatus", WWSUstatus)
 	.add("WWSUeas", WWSUeas)
@@ -98,7 +98,7 @@ wwsumodules
 	.add("WWSUrecipients", WWSUrecipients)
 	.add("WWSUhosts", WWSUhosts, {
 		machineID: machineID,
-		app: window.ipc.getAppVersion()
+		app: window.ipc.getAppVersion(),
 	})
 	.add("WWSUrequests", WWSUrequests)
 	.add("WWSUtimesheet", WWSUtimesheet)
@@ -153,10 +153,10 @@ let sounds = {
 	onBreak: new Howl({ src: ["assets/voice-queues/break.mp3"] }),
 	oneMinute: new Howl({ src: ["assets/voice-queues/oneMinute.mp3"] }),
 	thirtySeconds: new Howl({
-		src: ["assets/voice-queues/thirtySeconds.mp3"]
+		src: ["assets/voice-queues/thirtySeconds.mp3"],
 	}),
 	fifteenSeconds: new Howl({
-		src: ["assets/voice-queues/fifteenSeconds.mp3"]
+		src: ["assets/voice-queues/fifteenSeconds.mp3"],
 	}),
 	tenSeconds: new Howl({ src: ["assets/voice-queues/tenSeconds.mp3"] }),
 	fiveSeconds: new Howl({ src: ["assets/voice-queues/fiveSeconds.mp3"] }),
@@ -166,54 +166,56 @@ let sounds = {
 	callQuality: new Howl({ src: ["assets/voice-queues/callQuality.mp3"] }),
 	callSilence: new Howl({ src: ["assets/voice-queues/callSilence.mp3"] }),
 	callTerminated: new Howl({
-		src: ["assets/voice-queues/callTerminated.mp3"]
-	})
+		src: ["assets/voice-queues/callTerminated.mp3"],
+	}),
 };
 
 // letiables
 let breakNotified = false; // Did we notify the DJ they have to take a top of hour ID break?
-
 let queueLength = 0; // Current queue in seconds
-
 let countDown = 0; // Countdown to on air in seconds
 
+// Available audio devices
 let audioDevices = [];
 
 // Remote calls
 let pendingHostCall;
-
 let badQualityTimer;
 
+// TODO admin item
 let todos = {
 	status: {
 		danger: 0,
 		orange: 0,
 		warning: 0,
 		info: 0,
-		primary: 0
+		primary: 0,
 	},
 	accountability: {
 		danger: 0,
 		orange: 0,
 		warning: 0,
 		info: 0,
-		primary: 0
+		primary: 0,
 	},
 	timesheets: {
 		danger: 0,
 		orange: 0,
 		warning: 0,
 		info: 0,
-		primary: 0
+		primary: 0,
 	},
 	DJs: {
 		danger: 0,
 		orange: 0,
 		warning: 0,
 		info: 0,
-		primary: 0
-	}
+		primary: 0,
+	},
 };
+
+// What we notified as should be playing now
+let notifiedShouldBePlaying = ``;
 
 // Navigation
 navigation
@@ -279,14 +281,14 @@ navigation
 					properties: {
 						delay: {
 							type: "number",
-							title: "Delay (milliseconds)"
+							title: "Delay (milliseconds)",
 						},
 						recordPath: {
 							type: "string",
 							format: "uri",
-							title: "Path to audio recordings"
-						}
-					}
+							title: "Path to audio recordings",
+						},
+					},
 				},
 				options: {
 					fields: {
@@ -294,25 +296,25 @@ navigation
 							helper:
 								"How much time passes between a state change and when the input device receives the audio? For example, if the input device is subject to a delay system, you would put the amount of delay time in here.",
 							events: {
-								change: function() {
+								change: function () {
 									let value = this.getValue();
 									if (!this.handleValidate()) {
 										console.log(`invalid`);
 										return;
 									}
 									window.saveSettings.recorder("delay", value);
-								}
-							}
+								},
+							},
 						},
 						// TODO: re-compile alpaca with the capability of choosing a folder.
 						recordPath: {
 							helpers: [
 								`Write the full path to the directory you want audio files (webm format) to be saved`,
 								`Sub-directories for automation, remote, live, and sports will be created automatically after the first recording is saved.`,
-								`Additional sub-sub-directories will be created automatically to organize recordings by genre, show, or sport.`
+								`Additional sub-sub-directories will be created automatically to organize recordings by genre, show, or sport.`,
 							],
 							events: {
-								change: function() {
+								change: function () {
 									let value = this.getValue();
 									if (!this.handleValidate()) {
 										console.log(`invalid`);
@@ -320,12 +322,12 @@ navigation
 									}
 									window.saveSettings.recorder("recordPath", value);
 									startRecording(0);
-								}
-							}
-						}
-					}
+								},
+							},
+						},
+					},
 				},
-				data: window.settings.recorder()
+				data: window.settings.recorder(),
 			});
 			$("#section-audio-silence-form").alpaca({
 				schema: {
@@ -335,16 +337,16 @@ navigation
 							type: "number",
 							title: "Delay (milliseconds)",
 							required: true,
-							minimum: 0
+							minimum: 0,
 						},
 						threshold: {
 							type: "number",
 							title: "Threshold (percentile 0.0 - 1.0)",
 							minimum: 0,
 							maximum: 1,
-							required: true
-						}
-					}
+							required: true,
+						},
+					},
 				},
 				options: {
 					fields: {
@@ -352,7 +354,7 @@ navigation
 							helper:
 								"How much time should elapse when the volume of the combined input devices for silence monitoring drops below the threshold before DJ Controls triggers the silence alarm in WWSU?",
 							events: {
-								change: function() {
+								change: function () {
 									let value = this.getValue();
 									if (!this.handleValidate()) {
 										console.log(`invalid`);
@@ -360,14 +362,14 @@ navigation
 									}
 									window.saveSettings.silence("delay", value);
 									window.ipc.silence.setting([]);
-								}
-							}
+								},
+							},
 						},
 						threshold: {
 							helper:
 								"At what volume percentile (0.0 - 1.0) should silence be considered detected when the combined volumes of input devices with silence checked drop below this value?",
 							events: {
-								change: function() {
+								change: function () {
 									let value = this.getValue();
 									if (!this.handleValidate()) {
 										console.log(`invalid`);
@@ -375,12 +377,12 @@ navigation
 									}
 									window.saveSettings.silence("threshold", value);
 									window.ipc.silence.setting([]);
-								}
-							}
-						}
-					}
+								},
+							},
+						},
+					},
 				},
-				data: window.settings.silence()
+				data: window.settings.silence(),
 			});
 		}
 	)
@@ -522,7 +524,7 @@ $("#section-audio-devices-refresh").on("click", () => {
 		timeout: 30000,
 		onBlock: () => {
 			window.ipc.audioRefreshDevices(true);
-		}
+		},
 	});
 });
 $("#section-serial-delay-refresh").on("click", () => {
@@ -532,7 +534,7 @@ $("#section-serial-delay-refresh").on("click", () => {
 		timeout: 30000,
 		onBlock: () => {
 			refreshSerialPorts();
-		}
+		},
 	});
 });
 $(".chat-mute").on("click", () => {
@@ -551,7 +553,7 @@ $(".btn-dashboard-meta-clear").on("click", () => {
 				logsubtype: meta.meta ? meta.meta.show : "",
 				loglevel: "secondary",
 				logIcon: "fas fa-file",
-				title: "DJ / Producer began talking."
+				title: "DJ / Producer began talking.",
 			},
 			true
 		);
@@ -570,14 +572,14 @@ $(".btn-operation-resume").on("click", () => {
 		let called = recipients
 			.db()
 			.get()
-			.find(rec => rec.hostID === meta.meta.hostCalled);
+			.find((rec) => rec.hostID === meta.meta.hostCalled);
 		if (!called || !called.peer || called.status !== 5) {
 			$(document).Toasts("create", {
 				class: "bg-warning",
 				title: "Remote host not connected",
 				delay: 30000,
 				autohide: true,
-				body: `The host receiving the audio for the broadcast is not connected. Please wait until the resume button flashes to indicate it re-connected, or end and restart the remote broadcast with a different host.`
+				body: `The host receiving the audio for the broadcast is not connected. Please wait until the resume button flashes to indicate it re-connected, or end and restart the remote broadcast with a different host.`,
 			});
 			remote.request({ ID: meta.meta.hostCalled || pendingHostCall });
 			return;
@@ -658,43 +660,41 @@ climacell.initClockForecast("#weather-forecast-donut");
 // CLOCKWHEEL
 
 // Initialize clockwheel
-let clockwheelDonutCanvas = $("#clockwheel-donut")
-	.get(0)
-	.getContext("2d");
+let clockwheelDonutCanvas = $("#clockwheel-donut").get(0).getContext("2d");
 let clockwheelDonutData = {
 	labels: ["Not Yet Loaded"],
 	datasets: [
 		{
 			data: [60],
-			backgroundColor: ["#000000"]
+			backgroundColor: ["#000000"],
 		},
 		{
 			data: [720],
-			backgroundColor: ["#000000"]
-		}
-	]
+			backgroundColor: ["#000000"],
+		},
+	],
 };
 let clockwheelDonutOptions = {
 	maintainAspectRatio: false,
 	responsive: true,
 	cutoutPercentage: 66,
 	legend: {
-		display: false
+		display: false,
 	},
 	animation: {
 		animateRotate: false,
-		animateScale: false
+		animateScale: false,
 	},
 	elements: {
 		arc: {
-			borderColor: "rgba(0, 0, 0, 0)"
-		}
-	}
+			borderColor: "rgba(0, 0, 0, 0)",
+		},
+	},
 };
 let clockwheelDonut = new Chart(clockwheelDonutCanvas, {
 	type: "doughnut",
 	data: clockwheelDonutData,
-	options: clockwheelDonutOptions
+	options: clockwheelDonutOptions,
 });
 
 // Clockwheel Clock and functions
@@ -828,7 +828,7 @@ function startRecording(delay) {
 				`${recordState}/${preText}/${preText2} (${moment().format(
 					"YYYY_MM_DD HH_mm_ss"
 				)})`,
-				delay
+				delay,
 			]);
 		} else {
 			window.ipc.recorder.stop([-1]);
@@ -845,7 +845,7 @@ let fullCalendar = new FullCalendar.Calendar(calendarEl, {
 	headerToolbar: {
 		start: "prev,next today",
 		center: "title",
-		end: "dayGridMonth,timeGridWeek,timeGridDay,listWeek"
+		end: "dayGridMonth,timeGridWeek,timeGridDay,listWeek",
 	},
 	initialView: "timeGridWeek",
 	navLinks: true, // can click day/week names to navigate views
@@ -857,7 +857,7 @@ let fullCalendar = new FullCalendar.Calendar(calendarEl, {
 	themeSystem: "bootstrap",
 	dayMaxEvents: 5,
 	slotDuration: "00:15:00",
-	events: function(info, successCallback, failureCallback) {
+	events: function (info, successCallback, failureCallback) {
 		animations.add("calendar-update", () => {
 			$("#calendar").block({
 				message: "<h1>Loading...</h1>",
@@ -865,9 +865,9 @@ let fullCalendar = new FullCalendar.Calendar(calendarEl, {
 				timeout: 30000,
 				onBlock: () => {
 					calendar.getEvents(
-						events => {
+						(events) => {
 							events = events
-								.filter(event => {
+								.filter((event) => {
 									// Filter out events by filters
 									if (event.scheduleType === "canceled-changed") return false;
 									let temp = document.getElementById(`filter-${event.type}`);
@@ -877,7 +877,7 @@ let fullCalendar = new FullCalendar.Calendar(calendarEl, {
 										return false;
 									}
 								})
-								.map(event => {
+								.map((event) => {
 									let borderColor;
 									let title = `${event.type}: ${event.hosts} - ${event.name}`;
 									if (
@@ -918,33 +918,31 @@ let fullCalendar = new FullCalendar.Calendar(calendarEl, {
 											: "#e6e6e6",
 										borderColor: borderColor,
 										extendedProps: {
-											event: event
-										}
+											event: event,
+										},
 									};
 								});
 							successCallback(events);
 							fullCalendar.updateSize();
 							$("#calendar").unblock();
 						},
-						moment(info.start)
-							.subtract(1, "days")
-							.toISOString(true),
+						moment(info.start).subtract(1, "days").toISOString(true),
 						moment(info.end).toISOString(true)
 					);
-				}
+				},
 			});
 		});
 	},
 
-	eventClick: function(info) {
+	eventClick: function (info) {
 		calendar.showClickedEvent(info.event.extendedProps.event);
 	},
 
-	select: function(info) {
+	select: function (info) {
 		calendar.newOccurrence(info.startStr, info.endStr);
 	},
 
-	eventDrop: function(info) {
+	eventDrop: function (info) {
 		let duration = moment(info.event.end).diff(info.event.start, "minutes");
 		if (duration > 60 * 24) {
 			$(document).Toasts("create", {
@@ -953,7 +951,7 @@ let fullCalendar = new FullCalendar.Calendar(calendarEl, {
 				body:
 					"Occurrences may not last more than 24 hours. Consider setting up a recurring schedule.",
 				autohide: true,
-				delay: 15000
+				delay: 15000,
 			});
 			return;
 		}
@@ -965,7 +963,7 @@ let fullCalendar = new FullCalendar.Calendar(calendarEl, {
 		info.revert();
 	},
 
-	eventResize: function(info) {
+	eventResize: function (info) {
 		let duration = moment(info.event.end).diff(info.event.start, "minutes");
 		if (duration > 60 * 24) {
 			$(document).Toasts("create", {
@@ -974,7 +972,7 @@ let fullCalendar = new FullCalendar.Calendar(calendarEl, {
 				body:
 					"Occurrences may not last more than 24 hours. Consider setting up a recurring schedule.",
 				autohide: true,
-				delay: 15000
+				delay: 15000,
 			});
 			return;
 		}
@@ -984,7 +982,7 @@ let fullCalendar = new FullCalendar.Calendar(calendarEl, {
 			duration
 		);
 		info.revert();
-	}
+	},
 });
 fullCalendar.render();
 
@@ -999,31 +997,31 @@ fullCalendar.render();
 	"event",
 	"onair-booking",
 	"prod-booking",
-	"office-hours"
-].map(type => {
+	"office-hours",
+].map((type) => {
 	let temp = document.getElementById(`filter-${type}`);
 	if (temp !== null) {
-		temp.addEventListener("click", e => {
+		temp.addEventListener("click", (e) => {
 			fullCalendar.refetchEvents();
 		});
 	}
 });
 
 // Add click events to filter group buttons
-$("#filter-group-broadcasts").on("click", e => {
+$("#filter-group-broadcasts").on("click", (e) => {
 	[
 		"genre",
 		"event",
 		"onair-booking",
 		"prod-booking",
-		"office-hours"
-	].map(type => $(`#filter-${type}`).prop("checked", false));
-	["show", "sports", "remote", "prerecord", "playlist"].map(type =>
+		"office-hours",
+	].map((type) => $(`#filter-${type}`).prop("checked", false));
+	["show", "sports", "remote", "prerecord", "playlist"].map((type) =>
 		$(`#filter-${type}`).prop("checked", true)
 	);
 	fullCalendar.refetchEvents();
 });
-$("#filter-group-bookings").on("click", e => {
+$("#filter-group-bookings").on("click", (e) => {
 	[
 		"show",
 		"sports",
@@ -1032,14 +1030,14 @@ $("#filter-group-bookings").on("click", e => {
 		"genre",
 		"playlist",
 		"event",
-		"office-hours"
-	].map(type => $(`#filter-${type}`).prop("checked", false));
-	["onair-booking", "prod-booking"].map(type =>
+		"office-hours",
+	].map((type) => $(`#filter-${type}`).prop("checked", false));
+	["onair-booking", "prod-booking"].map((type) =>
 		$(`#filter-${type}`).prop("checked", true)
 	);
 	fullCalendar.refetchEvents();
 });
-$("#filter-group-clear").on("click", e => {
+$("#filter-group-clear").on("click", (e) => {
 	[
 		"show",
 		"sports",
@@ -1050,8 +1048,8 @@ $("#filter-group-clear").on("click", e => {
 		"event",
 		"office-hours",
 		"onair-booking",
-		"prod-booking"
-	].map(type => $(`#filter-${type}`).prop("checked", false));
+		"prod-booking",
+	].map((type) => $(`#filter-${type}`).prop("checked", false));
 	fullCalendar.refetchEvents();
 });
 
@@ -1111,13 +1109,13 @@ window.ipc.on.recorderFailed((event, arg) => {
 			loglevel: "orange",
 			logIcon: `fas fa-file-audio`,
 			title: `A recording failed!`,
-			event: `File: ${arg[0]}<br />Error: ${arg[1]}`
+			event: `File: ${arg[0]}<br />Error: ${arg[1]}`,
 		},
 		true
 	);
 	status.recorder({
 		status: 2,
-		data: `There was an error saving the file ${arg[0]}<br />Error: ${arg[1]}<br /><strong>Be prepared to record your broadcasts manually</strong> until the recorder is fixed.`
+		data: `There was an error saving the file ${arg[0]}<br />Error: ${arg[1]}<br /><strong>Be prepared to record your broadcasts manually</strong> until the recorder is fixed.`,
 	});
 });
 
@@ -1168,7 +1166,7 @@ window.ipc.on.silenceState((event, arg) => {
 
 // Update VU meters
 window.ipc.on.audioVolume((event, arg) => {
-	(volumes => {
+	((volumes) => {
 		if (navigation.activeMenu !== `#nav-audio`) {
 			return;
 		}
@@ -1199,13 +1197,13 @@ window.ipc.on.recorderSaved((event, arg) => {
 			loglevel: "info",
 			logIcon: `fas fa-file-audio`,
 			title: `A recording was saved.`,
-			event: `Path: ${arg}`
+			event: `Path: ${arg}`,
 		},
 		true
 	);
 	status.recorder({
 		status: 5,
-		data: `Most recent file ${arg} was successfully saved.`
+		data: `Most recent file ${arg} was successfully saved.`,
 	});
 });
 
@@ -1303,7 +1301,7 @@ window.ipc.on.audioDevices((event, arg) => {
 							100 * (5 / 8),
 							100 * (6 / 8),
 							100 * (7 / 8),
-							100
+							100,
 						],
 						ticks_labels: [
 							"OFF",
@@ -1314,56 +1312,59 @@ window.ipc.on.audioDevices((event, arg) => {
 							"125%",
 							"150%",
 							"175%",
-							"200%"
+							"200%",
 						],
 						ticks_snap_bounds: 0.025,
 						value: device.settings.volume,
 						orientation: "horizontal",
 						selection: "before",
-						tooltip: "show"
+						tooltip: "show",
 					});
 
 					// Volume slider listener
 					$(`#audio-volume-input-${device.device.deviceId}`).off("change");
 					$(`#audio-volume-input-${device.device.deviceId}`).on(
 						"change",
-						obj => {
+						(obj) => {
 							window.ipc.audioChangeVolume([
 								device.device.deviceId,
 								"audioinput",
-								obj.value.newValue
+								obj.value.newValue,
 							]);
 						}
 					);
 
 					// Checkbox listeners
 					$(`#audio-remote-input-${device.device.deviceId}`).off("change");
-					$(`#audio-remote-input-${device.device.deviceId}`).on("change", e => {
-						window.ipc.audioRemoteSetting([
-							device.device.deviceId,
-							"audioinput",
-							e.target.checked
-						]);
-					});
+					$(`#audio-remote-input-${device.device.deviceId}`).on(
+						"change",
+						(e) => {
+							window.ipc.audioRemoteSetting([
+								device.device.deviceId,
+								"audioinput",
+								e.target.checked,
+							]);
+						}
+					);
 					$(`#audio-recorder-input-${device.device.deviceId}`).off("change");
 					$(`#audio-recorder-input-${device.device.deviceId}`).on(
 						"change",
-						e => {
+						(e) => {
 							window.ipc.audioRecorderSetting([
 								device.device.deviceId,
 								"audioinput",
-								e.target.checked
+								e.target.checked,
 							]);
 						}
 					);
 					$(`#audio-silence-input-${device.device.deviceId}`).off("change");
 					$(`#audio-silence-input-${device.device.deviceId}`).on(
 						"change",
-						e => {
+						(e) => {
 							window.ipc.audioSilenceSetting([
 								device.device.deviceId,
 								"audioinput",
-								e.target.checked
+								e.target.checked,
 							]);
 						}
 					);
@@ -1408,32 +1409,32 @@ window.ipc.on.audioDevices((event, arg) => {
 							100 * (1 / 4),
 							100 * (2 / 4),
 							100 * (3 / 4),
-							100 * (4 / 4)
+							100 * (4 / 4),
 						],
 						ticks_labels: ["OFF", "25%", "50%", "75%", "100%"],
 						ticks_snap_bounds: 0.025,
 						value: device.settings.volume,
 						orientation: "horizontal",
 						selection: "before",
-						tooltip: "show"
+						tooltip: "show",
 					});
 
 					// Volume slider listener
 					$(`#audio-volume-output-${device.device.deviceId}`).off("change");
 					$(`#audio-volume-output-${device.device.deviceId}`).on(
 						"change",
-						obj => {
+						(obj) => {
 							window.ipc.audioChangeVolume([
 								device.device.deviceId,
 								"audiooutput",
-								obj.value.newValue
+								obj.value.newValue,
 							]);
 						}
 					);
 
 					// Checkbox listeners
 					$(`#audio-output-${device.device.deviceId}`).off("change");
-					$(`#audio-output-${device.device.deviceId}`).on("change", e => {
+					$(`#audio-output-${device.device.deviceId}`).on("change", (e) => {
 						if ($(`#audio-output-${device.device.deviceId}`).prop("checked")) {
 							$(`.form-check-devices-output`).each((index2, element) => {
 								if (element.id !== `audio-output-${device.device.deviceId}`) {
@@ -1443,30 +1444,36 @@ window.ipc.on.audioDevices((event, arg) => {
 							window.ipc.audioOutputSetting([
 								device.device.deviceId,
 								"audiooutput",
-								true
+								true,
 							]);
 						}
 					});
 
 					$(`#audio-queue-output-${device.device.deviceId}`).off("change");
-					$(`#audio-queue-output-${device.device.deviceId}`).on("change", e => {
-						if (
-							$(`#audio-queue-output-${device.device.deviceId}`).prop("checked")
-						) {
-							$(`.form-check-devices-queue`).each((index2, element) => {
-								if (
-									element.id !== `audio-queue-output-${device.device.deviceId}`
-								) {
-									$(element).prop({ checked: false });
-								}
-							});
-							window.ipc.audioQueueSetting([
-								device.device.deviceId,
-								"audiooutput",
-								true
-							]);
+					$(`#audio-queue-output-${device.device.deviceId}`).on(
+						"change",
+						(e) => {
+							if (
+								$(`#audio-queue-output-${device.device.deviceId}`).prop(
+									"checked"
+								)
+							) {
+								$(`.form-check-devices-queue`).each((index2, element) => {
+									if (
+										element.id !==
+										`audio-queue-output-${device.device.deviceId}`
+									) {
+										$(element).prop({ checked: false });
+									}
+								});
+								window.ipc.audioQueueSetting([
+									device.device.deviceId,
+									"audiooutput",
+									true,
+								]);
+							}
 						}
-					});
+					);
 				});
 			}
 		});
@@ -1550,7 +1557,6 @@ window.ipc.on.delay((event, args) => {
 // DISABLED: Does not work right now
 function refreshSerialPorts() {
 	// Get available ports (via custom Electron event in index.js)
-
 	// TODO: this is temporary until Electron fixes their user gesture bug
 	// window.ipc.getSerialPorts();
 }
@@ -1567,7 +1573,7 @@ window.ipc.on.serialPorts((event, ports) => {
 	let delayPorts = `<option value="">(NONE)</option>`;
 	ports = ports[0];
 	if (ports.constructor === Array && ports.length > 0) {
-		ports.map(port => {
+		ports.map((port) => {
 			delayPorts += `<option value="${port.deviceInstanceId}">${port.displayName} (${port.portName})</option>`;
 		});
 	}
@@ -1583,7 +1589,7 @@ window.ipc.on.serialPorts((event, ports) => {
 
 	// Add select box change handler for setting new port when changed
 	$("#section-serial-delay-port").unbind("change");
-	$("#section-serial-delay-port").on("change", e => {
+	$("#section-serial-delay-port").on("change", (e) => {
 		let val = $(e.target).val();
 		window.saveSettings.delay("port", val);
 		// Restart delay system by closing the process so we can use the new port
@@ -1605,7 +1611,7 @@ socket.on("connect", () => {
 	socket._raw.io._reconnectionAttempts = Infinity;
 
 	discipline.checkDiscipline(() => {
-		hosts.get(success => {
+		hosts.get((success) => {
 			if (success === 1) {
 				config.init();
 				meta.init();
@@ -1693,7 +1699,7 @@ socket.on("disconnect", () => {
 });
 
 // Connection error
-socket.on("reconnect_failed", error => {
+socket.on("reconnect_failed", (error) => {
 	$("#unauthorized").removeClass("d-none");
 	$("#connecting").addClass("d-none");
 	$("#reconnecting").addClass("d-none");
@@ -1718,7 +1724,7 @@ socket.on("delay-system-dump", () => {
 	}
 });
 
-animations.on("updateStatus", "renderer", updating => {
+animations.on("updateStatus", "renderer", (updating) => {
 	if (updating) {
 		$("#animation-refreshing").removeClass("d-none");
 	} else {
@@ -1740,7 +1746,7 @@ meta.on("newMeta", "renderer", (updated, fullMeta) => {
 					$(".operations").block({
 						message: `<h4>${fullMeta.changingState}</h4>`,
 						css: { border: "3px solid #a00" },
-						timeout: 60000
+						timeout: 60000,
 					});
 				} else {
 					$(".operations").unblock();
@@ -1986,7 +1992,7 @@ meta.on("newMeta", "renderer", (updated, fullMeta) => {
 				"sports_on",
 				"sportsremote_on",
 				"sports_returning",
-				"sportsremote_returning"
+				"sportsremote_returning",
 			].indexOf(updated.state) === -1
 		) {
 			window.ipc.process.remote(["close"]);
@@ -2029,13 +2035,13 @@ meta.on("newMeta", "renderer", (updated, fullMeta) => {
 				"There was an error in meta.newMeta. Please report this to the engineer.",
 			autohide: true,
 			delay: 10000,
-			icon: "fas fa-skull-crossbones fa-lg"
+			icon: "fas fa-skull-crossbones fa-lg",
 		});
 	}
 });
 
 // Meta ticker
-meta.on("metaTick", "renderer", fullMeta => {
+meta.on("metaTick", "renderer", (fullMeta) => {
 	try {
 		// Calculate queue time and countdown time
 		queueLength =
@@ -2131,7 +2137,7 @@ meta.on("metaTick", "renderer", fullMeta => {
 			) {
 				$(".operations-bar").removeClass("navbar-gray-dark");
 				$(".operations-bar").addClass("navbar-orange");
-				setTimeout(function() {
+				setTimeout(function () {
 					$(".operations-bar").removeClass("navbar-orange");
 					$(".operations-bar").addClass("navbar-gray-dark");
 				}, 500);
@@ -2293,8 +2299,8 @@ meta.on("metaTick", "renderer", fullMeta => {
 						header: "Do Not Forget the Top of the Hour Break!",
 						flash: true,
 						body:
-							"You are required to take a break before :05 past the hour unless you are ending your broadcast before then."
-					}
+							"You are required to take a break before :05 past the hour unless you are ending your broadcast before then.",
+					},
 				]);
 				window.ipc.flashMain(true);
 			}
@@ -2326,7 +2332,7 @@ meta.on("metaTick", "renderer", fullMeta => {
 				"There was an error in meta.metaTick. Please report this to the engineer.",
 			autohide: true,
 			delay: 10000,
-			icon: "fas fa-skull-crossbones fa-lg"
+			icon: "fas fa-skull-crossbones fa-lg",
 		});
 	}
 });
@@ -2350,15 +2356,19 @@ function processStatus(db) {
 		orange: 0,
 		warning: 0,
 		info: 0,
-		primary: 0
+		primary: 0,
 	};
 	recountTodos();
-	db.filter(record => record.status <= 4)
+	db.filter((record) => record.status <= 4)
 		.sort((a, b) => a.status - b.status)
-		.map(record => {
+		.map((record) => {
 			// Notifications on silence detection
 			// TODO: Better way to differentiate between actual silence detected and silence alarm not working
-			if (record.name === `silence` && hosts.isHost && record.data.includes("Silence / very low audio detected")) {
+			if (
+				record.name === `silence` &&
+				hosts.isHost &&
+				record.data.includes("Silence / very low audio detected")
+			) {
 				window.ipc.makeNotification([
 					{
 						title: "Silence Detected",
@@ -2366,8 +2376,8 @@ function processStatus(db) {
 						header:
 							"Silence detection triggered! Please ensure your audio levels are good and not too quiet.",
 						flash: true,
-						body: record.data
-					}
+						body: record.data,
+					},
 				]);
 			}
 
@@ -2442,7 +2452,7 @@ function processStatus(db) {
 	});
 }
 
-status.on("change", "renderer", db => {
+status.on("change", "renderer", (db) => {
 	processStatus(db.get());
 });
 
@@ -2460,7 +2470,7 @@ function processEas(db) {
 
 	// Process each status and generate content
 	let html = `<ul>`;
-	db.map(record => {
+	db.map((record) => {
 		switch (record.severity) {
 			case "Extreme":
 				html += `<li>
@@ -2574,10 +2584,10 @@ function processEas(db) {
 	});
 }
 
-eas.on("change", "renderer", db => {
+eas.on("change", "renderer", (db) => {
 	processEas(db.get());
 });
-eas.on("newAlert", "renderer", record => {
+eas.on("newAlert", "renderer", (record) => {
 	if (record.severity === "Extreme") {
 		iziToast.show({
 			class: "flash-bg",
@@ -2603,7 +2613,7 @@ eas.on("newAlert", "renderer", record => {
 			overlayColor: "rgba(255, 0, 0, 0.33)",
 			zindex: 5000,
 			layout: 2,
-			maxWidth: 640
+			maxWidth: 640,
 		});
 	} else if (record.severity === "Severe") {
 		iziToast.show({
@@ -2629,7 +2639,7 @@ eas.on("newAlert", "renderer", record => {
 			overlay: true,
 			zindex: 4000,
 			layout: 2,
-			maxWidth: 640
+			maxWidth: 640,
 		});
 	}
 });
@@ -2648,11 +2658,11 @@ function processAnnouncements(db) {
 		// First, process djcontrols announcements for the dashboard
 		let html = ``;
 		db.filter(
-			record =>
+			(record) =>
 				record.type === "djcontrols" &&
 				moment(meta.meta.time).isSameOrAfter(record.starts) &&
 				moment(meta.meta.time).isBefore(record.expires)
-		).map(record => {
+		).map((record) => {
 			if (record.level === "trivial") record.level === "secondary";
 			html += `<div class="alert alert-${record.level}">
                 <h5>${record.title}</h5>
@@ -2663,7 +2673,7 @@ function processAnnouncements(db) {
 	});
 }
 
-announcements.on("change", "renderer", db => {
+announcements.on("change", "renderer", (db) => {
 	processAnnouncements(db.get());
 });
 
@@ -2679,8 +2689,77 @@ function updateCalendar() {
 function updateClockwheel() {
 	// Ask the calendar process to recalculate clockwheel segments
 	animations.add("clockwheel-update", () => {
-		calendar.getEvents(events => {
+		calendar.getEvents((events) => {
 			window.ipc.calendar.updateClockwheel([events, meta.meta]);
+		});
+	});
+
+	// Update and process notifications in the event shows run into each other
+	animations.add("calendar-notifications-update", () => {
+		// Skip if this DJ Controls is not hosting a broadcast right now
+		if (!hosts.isHost) return;
+
+		calendar.whatShouldBePlaying((events) => {
+			// Do not check for automation events or any canceled entries
+			events = events.filter(
+				(event) =>
+					[
+						"show",
+						"remote",
+						"sports",
+						"prerecord",
+					].indexOf(event.type) !== -1 &&
+					["canceled", "canceled-system", "canceled-changed"].indexOf(
+						event.scheduleType
+					) === -1
+			);
+
+			// Reset if there are no available events
+			if (!events || events.length === 0) {
+				notifiedShouldBePlaying = ``;
+				return;
+
+				// Check if one of these events are scheduled to go on in 2 or less minutes
+			} else {
+				let nowEvent = null;
+				let notifyEvent = null;
+
+				// Check each event on the schedule
+				for (let i in events) {
+					if (!Object.prototype.hasOwnProperty.call(events, i)) continue;
+
+					// If the event matches what is on the air now, note that in the nowEvent variable.
+					if (events[i].unique === meta.meta.calendarUnique) {
+						nowEvent = events[i];
+						continue;
+					}
+					// If this condition is met, this program is scheduled to be on the air shortly (within <2 minutes).
+					else if (
+						moment(meta.meta ? meta.meta.time : undefined).isSameOrAfter(
+							moment(events[i].start).subtract(2, "minutes")
+						) && (!notifyEvent || notifyEvent.priority < events[i].priority)
+					) {
+						notifyEvent = events[i];
+					}
+				}
+
+				// If notifyEvent is still null, there are no notifications to present. Exit.
+				if (!notifyEvent) return;
+
+				// If the event we should be notifying was already notified, don't notify again. Otherwise, notify if its priority is higher than the current event on the air.
+				if (notifyEvent.unique !== notifiedShouldBePlaying && (!nowEvent || nowEvent.priority < notifyEvent.priority)) {
+					notifiedShouldBePlaying = notifyEvent.unique;
+					window.ipc.makeNotification([
+						{
+							title: "Another broadcast is scheduled",
+							bg: "primary",
+							header: "Wrap up your broadcast; another is on the schedule!",
+							flash: true,
+							body: `<p>Please wrap up your broadcast.</p><p>${notifyEvent.type}: ${notifyEvent.hosts} - ${notifyEvent.name} is scheduled to be on the air at ${moment(notifyEvent.start).format("h:mm A")}.</p>`,
+						},
+					]);
+				}
+			}
 		});
 	});
 }
@@ -2717,7 +2796,7 @@ logs.on("count", "renderer", (danger, orange, warning, info) => {
 		orange,
 		warning,
 		info,
-		primary: 0
+		primary: 0,
 	};
 	recountTodos();
 });
@@ -2779,7 +2858,7 @@ function recountTodos() {
         TRACK REQUESTS FUNCTIONS
     */
 
-requests.on("trackRequested", "renderer", request => {
+requests.on("trackRequested", "renderer", (request) => {
 	$(document).Toasts("create", {
 		class: "bg-primary",
 		title: "Track Requested",
@@ -2788,7 +2867,7 @@ requests.on("trackRequested", "renderer", request => {
 		body: `A track was requested.<br />
 Track: <strong>${request.trackname}</strong>`,
 		icon: "fas fa-record-vinyl fa-lg",
-		position: "bottomRight"
+		position: "bottomRight",
 	});
 	window.ipc.flashMain(true);
 });
@@ -2797,7 +2876,7 @@ Track: <strong>${request.trackname}</strong>`,
         RECIPIENTS FUNCTIONS
     */
 
-recipients.on("change", "renderer", db => {
+recipients.on("change", "renderer", (db) => {
 	messages.updateRecipientsTable();
 
 	// If this host wants to make a call, and the host we want to call is online and has a peer, start a call.
@@ -2810,7 +2889,7 @@ recipients.on("change", "renderer", db => {
 			meta.meta.state === "automation_sportsremote" ||
 			pendingHostCall)
 	) {
-		let called = db.get().find(rec => rec.hostID === meta.meta.hostCalled);
+		let called = db.get().find((rec) => rec.hostID === meta.meta.hostCalled);
 		if (called && called.peer && called.status === 5) {
 			console.log(
 				`Host ${called.hostID} is ready to take the call. Asking remote process to start audio call if not already in one.`
@@ -2822,7 +2901,7 @@ recipients.on("change", "renderer", db => {
 		}
 	}
 });
-recipients.on("recipientChanged", "renderer", recipient => {
+recipients.on("recipientChanged", "renderer", (recipient) => {
 	messages.changeRecipient(recipient);
 });
 
@@ -2831,14 +2910,14 @@ recipients.on("recipientChanged", "renderer", recipient => {
     */
 
 messages.on("remove", "renderer", (query, db) => {
-	messages.read = messages.read.filter(value => value !== query);
-	messages.notified = messages.notified.filter(value => value !== query);
+	messages.read = messages.read.filter((value) => value !== query);
+	messages.notified = messages.notified.filter((value) => value !== query);
 });
-messages.on("change", "renderer", db => {
+messages.on("change", "renderer", (db) => {
 	messages.updateRecipient();
 	messages.updateRecipientsTable();
 });
-messages.on("newMessage", "renderer", message => {
+messages.on("newMessage", "renderer", (message) => {
 	window.ipc.flashMain(true);
 	$(document).Toasts("create", {
 		class: "bg-primary",
@@ -2847,7 +2926,7 @@ messages.on("newMessage", "renderer", message => {
 		delay: 30000,
 		body: `${message.message}<p><strong>To reply:</strong> Click "Messages / Chat" in the left menu and select the recipient.</p>`,
 		icon: "fas fa-comment fa-lg",
-		position: "bottomRight"
+		position: "bottomRight",
 	});
 });
 
@@ -2855,7 +2934,7 @@ messages.on("newMessage", "renderer", message => {
 		HOSTS FUNCTIONS
 	*/
 
-hosts.on("clientChanged", "renderer", newClient => {
+hosts.on("clientChanged", "renderer", (newClient) => {
 	// Refresh the socket
 	socket.disconnect();
 	setTimeout(() => {
@@ -2868,8 +2947,8 @@ hosts.on("clientChanged", "renderer", newClient => {
 	*/
 
 let newestVersion = ``;
-_version.on("change", "renderer", db => {
-	let record = db.get().find(rec => rec.app === `wwsu-dj-controls`);
+_version.on("change", "renderer", (db) => {
+	let record = db.get().find((rec) => rec.app === `wwsu-dj-controls`);
 	if (!record) return;
 	if (record.version !== newestVersion) {
 		newestVersion = record.version;
@@ -2893,8 +2972,8 @@ _version.on("change", "renderer", db => {
 				  <li>MacOS: You may need to run the .pkg file from Finder. Open Finder, browse to the downloaded .pkg, and hold down the control key while clicking on it. Click "open" in the menu item. The warning dialog should now have an open button, allowing you to run the installer.</li>
 				  <li>Windows (10): Run the exe installer. If Windows displays a warning, click "more info" to expose the "Run Anyway" button.</li>
 				  <li>Antivirus / Firewall: Some antiviruses or firewalls may block the installer or DJ Controls application since it is unsigned. Add them as trusted applications.</li>
-				  </ul>`
-				}
+				  </ul>`,
+				},
 			]);
 		}
 	}
@@ -2905,7 +2984,7 @@ _version.on("change", "renderer", db => {
 	*/
 
 // When a remote broadcast is requested, begin the process of starting up an audio call
-state.on("startRemote", "renderer", host => {
+state.on("startRemote", "renderer", (host) => {
 	console.log(`Requested remote broadcast with host ${host}`);
 
 	// Check to make sure the selected host is online. If not, bail.
@@ -2920,7 +2999,7 @@ state.on("startRemote", "renderer", host => {
 			title: "Remote broadcast failed",
 			delay: 15000,
 			autohide: true,
-			body: `The host you selected to call is not online. Please try using a different host.`
+			body: `The host you selected to call is not online. Please try using a different host.`,
 		});
 		window.ipc.process.remote(["close"]);
 		pendingHostCall = null;
@@ -2939,7 +3018,7 @@ window.ipc.on.remoteReady((event, arg) => {
 		`Remote process ready. Grabbing a Skyway.js credential from WWSU.`
 	);
 	$(".remote-start-status").html("Getting a Skyway.js credential");
-	remote.credentialComputer({}, credential => {
+	remote.credentialComputer({}, (credential) => {
 		console.dir(credential);
 		if (!credential) {
 			pendingHostCall = undefined;
@@ -2951,7 +3030,7 @@ window.ipc.on.remoteReady((event, arg) => {
 				title: "Remote broadcast failed",
 				delay: 15000,
 				autohide: true,
-				body: `There was a problem getting a skyway.js authorization credential. Please report this to the engineer.`
+				body: `There was a problem getting a skyway.js authorization credential. Please report this to the engineer.`,
 			});
 		} else {
 			console.log(
@@ -2961,7 +3040,7 @@ window.ipc.on.remoteReady((event, arg) => {
 			window.ipc.remote.peerCredential([
 				credential.peerId,
 				credential.apiKey,
-				credential.authToken
+				credential.authToken,
 			]);
 		}
 	});
@@ -2994,7 +3073,7 @@ window.ipc.on.remotePeerReady((event, arg) => {
 		let called = recipients
 			.db()
 			.get()
-			.find(rec => rec.hostID === pendingHostCall);
+			.find((rec) => rec.hostID === pendingHostCall);
 		if (
 			called &&
 			called.peer &&
@@ -3024,7 +3103,7 @@ window.ipc.on.remotePeerUnavailable((event, arg) => {
 		title: "Remote broadcast failed",
 		delay: 15000,
 		autohide: true,
-		body: `The host you selected to call did not answer the call. Please try using a different host.`
+		body: `The host you selected to call did not answer the call. Please try using a different host.`,
 	});
 	window.ipc.process.remote(["close"]);
 	pendingHostCall = null;
@@ -3035,7 +3114,7 @@ window.ipc.on.remoteIncomingCall((event, arg) => {
 	let recipient = recipients.find({
 		peer: arg[0],
 		makeCalls: true,
-		authorized: true
+		authorized: true,
 	});
 	if (recipient) {
 		console.log(
@@ -3066,7 +3145,7 @@ window.ipc.on.peerCallEstablished((event, arg) => {
 		!meta.meta.state.endsWith("_sportsremote")
 	) {
 		$(".remote-start-status").html("Starting remote broadcast");
-		state.finalizeRemote(success => {
+		state.finalizeRemote((success) => {
 			state.unblockBroadcastModal();
 			if (success) {
 				$(".remote-start-status").html("");
@@ -3089,7 +3168,7 @@ window.ipc.on.peerCallEstablished((event, arg) => {
 			title: "Remote call re-established",
 			delay: 30000,
 			autohide: true,
-			body: `The remote audio call was re-established. You can resume / proceed with the broadcast.`
+			body: `The remote audio call was re-established. You can resume / proceed with the broadcast.`,
 		});
 	}
 });
@@ -3129,8 +3208,8 @@ window.ipc.on.peerOutgoingSilence((event, arg) => {
 				bg: "danger",
 				header: "Silence on outgoing audio!",
 				flash: true,
-				body: `<p>Silence was detected for 15 seconds on outgoing audio. Remote broadcast has been sent to break. Please check your audio devices and DJ Controls' Audio settings.</p>`
-			}
+				body: `<p>Silence was detected for 15 seconds on outgoing audio. Remote broadcast has been sent to break. Please check your audio devices and DJ Controls' Audio settings.</p>`,
+			},
 		]);
 		sounds.callSilence.play();
 		window.ipc.process.remote(["close"]); // Restart process in case it is a process problem
@@ -3185,8 +3264,8 @@ window.ipc.on.peerCallClosed((event, arg) => {
 				bg: "danger",
 				header: "Audio Call was Closed",
 				flash: true,
-				body: `<p>The audio call for the remote broadcast closed. The broadcast was sent to break. Please check your network settings and resume the broadcast when things are stable.</p><p>This could also be a network issue on WWSU's end. If so, please report this under "report a problem".</p>`
-			}
+				body: `<p>The audio call for the remote broadcast closed. The broadcast was sent to break. Please check your network settings and resume the broadcast when things are stable.</p><p>This could also be a network issue on WWSU's end. If so, please report this under "report a problem".</p>`,
+			},
 		]);
 		sounds.callTerminated.play();
 
@@ -3223,8 +3302,8 @@ window.ipc.on.peerDestroyed((event, arg) => {
 				bg: "danger",
 				header: "Audio Call was Closed",
 				flash: true,
-				body: `<p>The audio call for the remote broadcast closed. The broadcast was sent to break. Please check your network settings and resume the broadcast when things are stable.</p><p>This could also be a network issue on WWSU's end. If so, please report this under "report a problem".</p>`
-			}
+				body: `<p>The audio call for the remote broadcast closed. The broadcast was sent to break. Please check your network settings and resume the broadcast when things are stable.</p><p>This could also be a network issue on WWSU's end. If so, please report this under "report a problem".</p>`,
+			},
 		]);
 		sounds.callTerminated.play();
 		pendingHostCall = undefined;
@@ -3281,8 +3360,8 @@ window.ipc.on.peerNoCalls((event, arg) => {
 				bg: "danger",
 				header: "Audio Call was Closed",
 				flash: true,
-				body: `<p>You tried to start / resume a broadcast when an audio call was not ongoing. The broadcast was sent to break. Please check your network settings and wait for the resume button to start pulsing.</p>`
-			}
+				body: `<p>You tried to start / resume a broadcast when an audio call was not ongoing. The broadcast was sent to break. Please check your network settings and wait for the resume button to start pulsing.</p>`,
+			},
 		]);
 		sounds.callTerminated.play();
 		pendingHostCall = undefined;
@@ -3322,7 +3401,7 @@ remoteQuality.on("quality", "renderer", (connection, reason, quality) => {
 	}
 });
 
-remote.on("callQuality", "renderer", quality => {
+remote.on("callQuality", "renderer", (quality) => {
 	if (
 		quality <= 0 &&
 		(meta.meta.state.startsWith("remote_") ||
@@ -3344,8 +3423,8 @@ remote.on("callQuality", "renderer", quality => {
 					bg: "warning",
 					header: "Poor Audio Call Quality",
 					flash: true,
-					body: `<p>Audio call quality is poor; call will be restarted the next time you take a break (unless it improves). Please check your network connection and ensure you are not running CPU-heavy programs.</p>`
-				}
+					body: `<p>Audio call quality is poor; call will be restarted the next time you take a break (unless it improves). Please check your network connection and ensure you are not running CPU-heavy programs.</p>`,
+				},
 			]);
 			badQualityTimer = setTimeout(() => {
 				badQualityTimer = undefined;